extract parseNLConversions for test
